+++ src/main/java/org/opensearch/flowframework/util/TenantAwareHelper.java
@@ -12,16 +12,22 @@ import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.rest.RestStatus;
 import org.opensearch.flowframework.common.CommonValue;
 import org.opensearch.flowframework.exception.FlowFrameworkException;
+import org.opensearch.flowframework.transport.WorkflowResponse;
 import org.opensearch.rest.RestRequest;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Helper class for tenant ID validation
  */
 public class TenantAwareHelper {
+
+    private static final ConcurrentHashMap<String, AtomicInteger> activeProvisionsPerTenant = new ConcurrentHashMap<>();
+    private static final ConcurrentHashMap<String, AtomicInteger> activeDeprovisionsPerTenant = new ConcurrentHashMap<>();
 
     private TenantAwareHelper() {}
 
@@ -87,5 +93,143 @@ public class TenantAwareHelper {
         }
 
         return tenantId;
+    }
+
+    /**
+     * Attempts to acquire a provision slot for the given tenant.
+     *
+     * @param maxExecutions The maximum number of simultaneous provisions allowed per tenant.
+     * @param tenantId The ID of the tenant requesting the provision.
+     * @param workflowListener The listener to notify in case of failure.
+     * @return true if the provision slot was acquired, false otherwise.
+     */
+    public static boolean tryAcquireProvision(int maxExecutions, String tenantId, ActionListener<WorkflowResponse> workflowListener) {
+        if (!tryAcquire(tenantId, activeProvisionsPerTenant, maxExecutions)) {
+            workflowListener.onFailure(
+                new FlowFrameworkException(
+                    "Exceeded max simultaneous provisioning requests: " + maxExecutions,
+                    RestStatus.TOO_MANY_REQUESTS
+                )
+            );
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Attempts to acquire a deprovision slot for the given tenant.
+     *
+     * @param maxExecutions The maximum number of simultaneous deprovisions allowed per tenant.
+     * @param tenantId The ID of the tenant requesting the deprovision.
+     * @param workflowListener The listener to notify in case of failure.
+     * @return true if the deprovision slot was acquired, false otherwise.
+     */
+    public static boolean tryAcquireDeprovision(int maxExecutions, String tenantId, ActionListener<WorkflowResponse> workflowListener) {
+        if (!tryAcquire(tenantId, activeDeprovisionsPerTenant, maxExecutions)) {
+            workflowListener.onFailure(
+                new FlowFrameworkException(
+                    "Exceeded max simultaneous deprovisioning requests: " + maxExecutions,
+                    RestStatus.TOO_MANY_REQUESTS
+                )
+            );
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * Releases a provision slot for the given tenant.
+     *
+     * @param tenantId The ID of the tenant for which to release the provision slot.
+     */
+    public static void releaseProvision(String tenantId) {
+        release(tenantId, activeProvisionsPerTenant);
+    }
+
+    /**
+     * Releases a deprovision slot for the given tenant.
+     *
+     * @param tenantId The ID of the tenant for which to release the deprovision slot.
+     */
+    public static void releaseDeprovision(String tenantId) {
+        release(tenantId, activeDeprovisionsPerTenant);
+    }
+
+    /**
+     * Attempts to acquire an execution slot for the given tenant.
+     *
+     * @param tenantId The ID of the tenant requesting the execution slot.
+     * @param executionsMap The map tracking the number of active executions per tenant.
+     * @param maxExecutions The maximum number of simultaneous executions allowed per tenant.
+     * @return true if the execution slot was acquired, false otherwise.
+     */
+    private static boolean tryAcquire(String tenantId, ConcurrentHashMap<String, AtomicInteger> executionsMap, int maxExecutions) {
+        if (tenantId == null) {
+            return true; // No throttling for null tenantId
+        }
+        AtomicInteger count = executionsMap.computeIfAbsent(tenantId, k -> new AtomicInteger(0));
+        if (count.incrementAndGet() <= maxExecutions) {
+            return true;
+        } else {
+            // If we're here we have exceeded maxExecutions, roll back
+            count.decrementAndGet();
+            return false;
+        }
+    }
+
+    /**
+     * Releases an execution slot for the given tenant.
+     *
+     * @param tenantId The ID of the tenant for which to release the execution slot.
+     * @param executionsMap The map tracking the number of active executions per tenant.
+     */
+    private static void release(String tenantId, ConcurrentHashMap<String, AtomicInteger> executionsMap) {
+        if (tenantId == null) {
+            return; // No throttling for null tenantId
+        }
+        executionsMap.computeIfPresent(tenantId, (key, count) -> {
+            int newValue = count.decrementAndGet();
+            // returning null from computeIfPresent removes from the map
+            return newValue > 0 ? count : null;
+        });
+    }
+
+    /**
+     * Create an action listener that releases provision throttle on failure only
+     * @param tenantId The tenant ID
+     * @param delegate the wrapped listener
+     * @return a listener wrapping the delegate that calls {@link #releaseProvision(String)} on failure of the wrapped listener.
+     */
+    public static ActionListener<WorkflowResponse> releaseProvisionOnFailureListener(
+        String tenantId,
+        ActionListener<WorkflowResponse> delegate
+    ) {
+        // Since this is only called on async failure we can't use runBefore, so we just copy its code in one case
+        return ActionListener.notifyOnce(new ActionListener<>() {
+            @Override
+            public void onResponse(WorkflowResponse response) {
+                delegate.onResponse(response);
+            }
+
+            @Override
+            public void onFailure(Exception e) {
+                try {
+                    releaseProvision(tenantId);
+                } finally {
+                    delegate.onFailure(e);
+                }
+            }
+        });
+    }
+
+    /**
+     * Create an action listener that releases deprovision throttle
+     * @param tenantId The tenant ID
+     * @param delegate the wrapped listener
+     * @return a listener wrapping the delegate that calls {@link #releaseDeprovision(String)} on the wrapped listener.
+     */
+    public static ActionListener<WorkflowResponse> releaseDeprovisionListener(String tenantId, ActionListener<WorkflowResponse> delegate) {
+        // Since this is only called on synchronous calls we release on both success and failure
+        return ActionListener.notifyOnce(ActionListener.runBefore(delegate, () -> releaseDeprovision(tenantId)));
     }
 }